Remove duplicated PopulateUI for stateful effects...
... And Effect no longer inherits wxEvtHandler!

And StatefulEffectBase is UI-agnostic
