+++ src/effects/StatefulEffect.cpp
@@ -12,7 +12,6 @@
 
 **********************************************************************/
 #include "StatefulEffect.h"
-#include "StatefulEffectUIServices.h"
 #include "SampleCount.h"
 #include "ShuttleGui.h"
 #include <wx/sizer.h>
@@ -43,28 +42,4 @@ std::shared_ptr<EffectInstance> StatefulEffect::MakeInstance() const
    // Stateless effects should override this function and be really const
    // correct.
    return std::make_shared<Instance>(const_cast<StatefulEffect&>(*this));
-}
-
-std::unique_ptr<EffectEditor> StatefulEffect::PopulateUI(
-   const EffectPlugin &, ShuttleGui &S,
-   EffectInstance &instance, EffectSettingsAccess &access,
-   const EffectOutputs *pOutputs) const
-{
-   auto parent = S.GetParent();
-
-   // As in MakeInstance, we still cheat const for stateful effects!
-   auto pThis = const_cast<StatefulEffect*>(this);
-
-   // Let the effect subclass provide its own editor if it wants
-   auto result = pThis->PopulateOrExchange(S, instance, access, pOutputs);
-
-   parent->SetMinSize(parent->GetSizer()->GetMinSize());
-
-   if (!result) {
-      // No custom editor object?  Then use the default
-      result = std::make_unique<DefaultEffectEditor>(*pThis,
-         *pThis, access, S.GetParent());
-      parent->PushEventHandler(pThis);
-   }
-   return result;
 }
