+++ src/effects/StatefulEffectBase.cpp
@@ -198,13 +198,6 @@ bool StatefulEffectBase::RealtimeFinalize(EffectSettings &settings) noexcept
    return false;
 }
 
-std::unique_ptr<EffectEditor> StatefulEffectBase::PopulateOrExchange(
-   ShuttleGui &, EffectInstance &, EffectSettingsAccess &,
-   const EffectOutputs *)
-{
-   return nullptr;
-}
-
 bool StatefulEffectBase::Init()
 {
    return true;
