+++ src/effects/StatefulPerTrackEffect.cpp
@@ -17,7 +17,6 @@
 
 *//*******************************************************************/
 #include "StatefulPerTrackEffect.h"
-#include "StatefulEffectUIServices.h"
 #include "ShuttleGui.h"
 #include <wx/sizer.h>
 
@@ -51,30 +50,6 @@ std::shared_ptr<EffectInstance> StatefulPerTrackEffect::MakeInstance() const
    // correct.
    return std::make_shared<Instance>(
       const_cast<StatefulPerTrackEffect&>(*this));
-}
-
-std::unique_ptr<EffectEditor>
-StatefulPerTrackEffect::PopulateUI(const EffectPlugin &, ShuttleGui &S,
-   EffectInstance &instance, EffectSettingsAccess &access,
-   const EffectOutputs *pOutputs) const
-{
-   auto parent = S.GetParent();
-
-   // As in MakeInstance, we still cheat const for stateful effects!
-   auto pThis = const_cast<StatefulPerTrackEffect*>(this);
-
-   // Let the effect subclass provide its own editor if it wants
-   auto result = pThis->PopulateOrExchange(S, instance, access, pOutputs);
-
-   parent->SetMinSize(parent->GetSizer()->GetMinSize());
-
-   if (!result) {
-      // No custom editor object?  Then use the default
-      result = std::make_unique<DefaultEffectEditor>(*pThis,
-         *pThis, access, S.GetParent());
-      parent->PushEventHandler(pThis);
-   }
-   return result;
 }
 
 bool StatefulPerTrackEffect::Process(
